+++ src/textual/widgets/__init__.py
@@ -9,7 +9,7 @@ from ..case import camel_to_snake
 # be able to "see" them.
 if typing.TYPE_CHECKING:
     from ._button import Button
-    from ._checkbox import Checkbox
+    from ._switch import Switch
     from ._data_table import DataTable
     from ._directory_tree import DirectoryTree
     from ._footer import Footer
@@ -29,7 +29,7 @@ if typing.TYPE_CHECKING:
 
 __all__ = [
     "Button",
-    "Checkbox",
+    "Switch",
     "DataTable",
     "DirectoryTree",
     "Footer",