+++ tests/test_focus.py
@@ -152,21 +152,21 @@ def test_focus_next_and_previous_with_type_selector_without_self():
     screen = app.screen
 
     from textual.containers import Horizontal, Vertical
-    from textual.widgets import Button, Checkbox, Input
+    from textual.widgets import Button, Switch, Input
 
     screen._add_children(
         Vertical(
             Horizontal(
                 Input(id="w3"),
-                Checkbox(id="w4"),
+                Switch(id="w4"),
                 Input(id="w5"),
                 Button(id="w6"),
-                Checkbox(id="w7"),
+                Switch(id="w7"),
                 id="w2",
             ),
             Horizontal(
                 Button(id="w9"),
-                Checkbox(id="w10"),
+                Switch(id="w10"),
                 Button(id="w11"),
                 Input(id="w12"),
                 Input(id="w13"),
@@ -180,11 +180,11 @@ def test_focus_next_and_previous_with_type_selector_without_self():
     assert screen.focused.id == "w3"
 
     assert screen.focus_next(Button).id == "w6"
-    assert screen.focus_next(Checkbox).id == "w7"
+    assert screen.focus_next(Switch).id == "w7"
     assert screen.focus_next(Input).id == "w12"
 
     assert screen.focus_previous(Button).id == "w11"
-    assert screen.focus_previous(Checkbox).id == "w10"
+    assert screen.focus_previous(Switch).id == "w10"
     assert screen.focus_previous(Button).id == "w9"
     assert screen.focus_previous(Input).id == "w5"